requirements.txt
GitPython
ipykernel
transormers
openai